+++ spec/support/shared_examples/authenticator.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-shared_examples MAuth::Client::AuthenticatorBase do
+shared_examples MAuth::Client::Authenticator do
   context 'when v2 and v1 headers are present on the object to authenticate' do
     it 'authenticates with v2' do
       signed_request = client.signed(request)
@@ -199,18 +199,16 @@ shared_examples MAuth::Client::AuthenticatorBase do
       end
     end
 
-    [::Faraday::ConnectionFailed, ::Faraday::TimeoutError].each do |error_klass|
+    [Faraday::ConnectionFailed, Faraday::TimeoutError].each do |error_klass|
       it "raises UnableToAuthenticate if mauth is unreachable with #{error_klass.name}" do
-        allow(test_faraday).to receive(:get).and_raise(error_klass.new('')) # for the local authenticator
-        allow(test_faraday).to receive(:post).and_raise(error_klass.new('')) # for the remote authenticator
+        allow(test_faraday).to receive(:get).and_raise(error_klass.new(''))
         expect { authenticating_mc.authentic?(v1_signed_req) }.to raise_error(MAuth::UnableToAuthenticateError)
       end
     end
 
     it 'raises UnableToAuthenticate if mauth errors' do
       stubs.instance_eval { @stack.clear } # HAX
-      stubs.get("/mauth/v1/security_tokens/#{app_uuid}.json") { [500, {}, []] } # for the local authenticator
-      stubs.post('/mauth/v1/authentication_tickets.json') { [500, {}, []] } # for the remote authenticator
+      stubs.get("/mauth/v1/security_tokens/#{app_uuid}.json") { [500, {}, []] }
       expect { authenticating_mc.authentic?(v1_signed_req) }.to raise_error(MAuth::UnableToAuthenticateError)
     end
 
@@ -221,8 +219,8 @@ shared_examples MAuth::Client::AuthenticatorBase do
 
       it 'logs the mauth app uuid of the requester and requestee when they both have such uuids' do
         expect(authenticating_mc.logger).to receive(:info).with(
-          'Mauth-client attempting to authenticate request from app with mauth app' \
-          ' uuid signer to app with mauth app uuid authenticator using version MWS.'
+          'Mauth-client attempting to authenticate request from app with mauth app ' \
+          'uuid signer to app with mauth app uuid authenticator using version MWS.'
         )
         authenticating_mc.authentic?(v1_signed_req)
       end
@@ -230,8 +228,8 @@ shared_examples MAuth::Client::AuthenticatorBase do
       it 'logs when the mauth app uuid is not provided in the request' do
         allow(v1_signed_req).to receive(:signature_app_uuid).and_return(nil)
         expect(authenticating_mc.logger).to receive(:info).with(
-          'Mauth-client attempting to authenticate request from app with mauth app' \
-          ' uuid [none provided] to app with mauth app uuid authenticator using version MWS.'
+          'Mauth-client attempting to authenticate request from app with mauth app ' \
+          'uuid [none provided] to app with mauth app uuid authenticator using version MWS.'
         )
         authenticating_mc.authentic?(v1_signed_req) rescue nil
       end
@@ -256,8 +254,8 @@ shared_examples MAuth::Client::AuthenticatorBase do
       signed_request.headers['MCC-Authentication'] = ''
       expect { authenticating_mc.authenticate!(signed_request) }.to raise_error(
         MAuth::MAuthNotPresent,
-        'Authentication Failed. No mAuth signature present; X-MWS-Authentication' \
-        ' header is blank, MCC-Authentication header is blank.'
+        'Authentication Failed. No mAuth signature present; X-MWS-Authentication ' \
+        'header is blank, MCC-Authentication header is blank.'
       )
     end
   end